// frontend/src/api/generated/models/TagDefinition.ts
export type TagDefinition = { type: 'integer' } & IntegerTagDefinition | { type: 'integerSet' } & IntegerSetTagDefinition | { type: 'keyOnly' } & KeyOnlyTagDefinition | { type: 'real' } & RealTagDefinition | { type: 'realSet' } & RealSetTagDefinition | { type: 'text' } & TextTagDefinition | { type: 'textSet' } & TextSetTagDefinition;